Created users edit and update user story

+++ db/seeds.rb
@@ -6,7 +6,30 @@
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
 require 'faker'
-Job.destroy_all
+
+puts "Destroy Jobs"
+Job.destroy_all if Rails.env.development?
+
+puts "Destroy Performers"
+User.destroy_all if Rails.env.development?
+
+puts "Create Performers"
+20.times do
+  attr = {
+    name: Faker::Name.name,
+    email: Faker::Internet.email,
+    password: "123123",
+    password_confirmation: "123123",
+    gender: Faker::Demographic.sex,
+    physical_attributes: Faker::Demographic.height(unit: :imperial),
+    ethnicity: Faker::Demographic.race,
+    age: Random.rand(18...42)
+  }
+  new_performer = User.new(attr)
+  new_performer.save
+end
+
+puts "Create Jobs"
 20.times do
   attr = {
     title: Faker::Job.title,
@@ -17,7 +40,9 @@ Job.destroy_all
     job_status: "Pro-Bono"
   }
   new_job = Job.new(attr)
-  my_user = User.find(1)
+  my_user = User.offset(rand(User.count)).first
   new_job.user = my_user
   new_job.save
 end
+
+puts "Complete Seed"
